various changes to get all tests passing via running pytest in the root maestro directory
