+++ maestro/cli/__init__.py
@@ -15,9 +15,3 @@
 import dotenv
 
 dotenv.load_dotenv()
-
-from .cli import CLI
-
-__all__ = {
-    "CLI"
-}
