+++ maestro/src/__init__.py
@@ -1,4 +1,25 @@
-# SPDX-License-Identifier: Apache-2.0
+# Copyright © 2025 IBM
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import dotenv
+
+dotenv.load_dotenv()
+
 from .workflow import Workflow
 from .bee_agent import BeeAgent
 from .agent import save_agent, restore_agent, remove_agent
+
+__all__ = {
+    "Workflow"
+}
