+++ maestro/tests/examples/test_condition.py
@@ -1,24 +1,18 @@
 #! /usr/bin/env python3
 # SPDX-License-Identifier: Apache-2.0
 
-import json
-import os
-import sys
+import os, sys, json, dotenv, yaml
 
-import dotenv
 from openai import OpenAI
-import yaml
-from maestro import Workflow
+
+from src.workflow import Workflow
 
 dotenv.load_dotenv()
-
 
 def parse_yaml(file_path):
     with open(file_path, "r") as file:
         yaml_data = list(yaml.safe_load_all(file))
     return yaml_data
-
-
 
 if __name__ == "__main__":
     agents_yaml = parse_yaml(os.path.join(os.path.dirname(__file__),"condition_agents.yaml"))
